+++ tests/testthat/EXS_collectRadii.R
@@ -1,5 +1,5 @@
-load("Scale_1_DHO.RData")
+load("testdata/Scale_1_DHO.RData")
 dat1 <- dat
 
-load("Oto140306_DHO.RData")
+load("testdata/Oto140306_DHO.RData")
 dat2 <- dat